# docs/openapi.yaml
openapi: 3.0.1
info:
  title: Bank Card Management API
  description: API для управления банковскими картами
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local server

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        role:
          type: string
          enum: [USER, ADMIN]
        createdAt:
          type: string
          format: date-time
    
    Card:
      type: object
      properties:
        id:
          type: integer
          format: int64
        cardNumber:
          type: string
        maskedCardNumber:
          type: string
        ownerName:
          type: string
        expiryDate:
          type: string
        status:
          type: string
          enum: [ACTIVE, BLOCKED, EXPIRED]
        balance:
          type: number
          format: double
        user:
          $ref: '#/components/schemas/User'
    
    AuthRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
        password:
          type: string
    
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        username:
          type: string
        role:
          type: string
    
    TransferRequest:
      type: object
      required:
        - fromCardId
        - toCardId
        - amount
      properties:
        fromCardId:
          type: integer
          format: int64
        toCardId:
          type: integer
          format: int64
        amount:
          type: number
          format: double

paths:
  /api/auth/register:
    post:
      tags:
        - Authentication
      summary: Регистрация нового пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequest'
      responses:
        200:
          description: Успешная регистрация
          content:
            text/plain:
              schema:
                type: string
        400:
          description: Ошибка регистрации

  /api/auth/login:
    post:
      tags:
        - Authentication
      summary: Вход в систему
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequest'
      responses:
        200:
          description: Успешный вход
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        401:
          description: Неверные учетные данные

  /api/cards:
    get:
      security:
        - bearerAuth: []
      tags:
        - Cards
      summary: Получить все карты
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 10
        - name: sortBy
          in: query
          schema:
            type: string
            default: id
      responses:
        200:
          description: Список карт
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/Card'
    
    post:
      security:
        - bearerAuth: []
      tags:
        - Cards
      summary: Создать новую карту
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Card'
      responses:
        200:
          description: Созданная карта
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Card'

  /api/cards/my:
    get:
      security:
        - bearerAuth: []
      tags:
        - Cards
      summary: Получить свои карты
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 10
      responses:
        200:
          description: Список карт пользователя

  /api/transfers:
    post:
      security:
        - bearerAuth: []
      tags:
        - Transfers
      summary: Перевод между картами
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferRequest'
      responses:
        200:
          description: Перевод выполнен
        400:
          description: Ошибка перевода

  /api/admin/users:
    get:
      security:
        - bearerAuth: []
      tags:
        - Admin
      summary: Получить всех пользователей
      responses:
        200:
          description: Список пользователей
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        403:
          description: Доступ запрещен

  /api/admin/cards:
    get:
      security:
        - bearerAuth: []
      tags:
        - Admin
      summary: Получить все карты (только ADMIN)
      responses:
        200:
          description: Список всех карт